Warning Cleanup and Maintainence

+++ AuxPort/Graphics/Images/AuxImage.cpp
@@ -17,7 +17,7 @@ void AuxPort::Graphics::GrayScaleImage::addColumn(const std::vector<float>& imag
 	AuxAssert(imageColumn.size() == height,"Image Column size does not match the height of the image");
 	auto imageChannel = imageData[columnNumber].data();
 	for (uint32_t i = 0;i < imageColumn.size();i++)
-		imageChannel[i] = AuxPort::Utility::remap<float>(imageColumn[i], 0, 1, minRange, maxRange);
+		imageChannel[i] = AuxPort::Utility::remap<float>(imageColumn[i], 0.0f, 1.0f, static_cast<float>(minRange), static_cast<float>(maxRange));
 }
 
 std::vector<std::vector<float>>* AuxPort::Graphics::GrayScaleImage::getImageData()
@@ -25,7 +25,7 @@ std::vector<std::vector<float>>* AuxPort::Graphics::GrayScaleImage::getImageData
 	return &imageData;
 }
 
-uint8_t AuxPort::Graphics::GrayScaleImage::getPoint(uint32_t x, uint32_t y) const noexcept
+float AuxPort::Graphics::GrayScaleImage::getPoint(uint32_t x, uint32_t y) const noexcept
 {
 	AuxAssert(x < imageData.size(), "Queried Row is greater than number of rows");
 	AuxAssert(y < imageData[0].size(), "Queried Column is greater than number of columns");
@@ -33,12 +33,12 @@ uint8_t AuxPort::Graphics::GrayScaleImage::getPoint(uint32_t x, uint32_t y) cons
 }
 
 
-int AuxPort::Graphics::GrayScaleImage::getRows() const noexcept
+size_t AuxPort::Graphics::GrayScaleImage::getRows() const noexcept
 {
 	return imageData.size();
 }
 
-int AuxPort::Graphics::GrayScaleImage::getColumns() const noexcept
+size_t AuxPort::Graphics::GrayScaleImage::getColumns() const noexcept
 {
 	return imageData[0].size();
 }
